feat: split navbar into multiple files
